keyguard/platforms/windows.py
# keyguard/platforms/windows.py
import win32api
import win32con
import win32security
import threading

class WindowsKeyboardHandler:
    def __init__(self):
        self.running = False
        self.hook = None
        
    def check_permissions(self) -> bool:
        try:
            security = win32security.OpenProcessToken(
                win32api.GetCurrentProcess(),
                win32con.TOKEN_QUERY
            )
            return True
        except:
            return False
            
    def start_protection(self):
        self.running = True
        self._protection_thread = threading.Thread(target=self._protect_keyboard)
        self._protection_thread.daemon = True
        self._protection_thread.start()
        
    def stop_protection(self):
        self.running = False
        
    def _protect_keyboard(self):
        while self.running:
            try:
                # Monitor keyboard device driver access
                handle = win32api.OpenProcess(
                    win32con.PROCESS_QUERY_INFORMATION,
                    False,
                    win32api.GetCurrentProcessId()
                )
                win32api.CloseHandle(handle)
            except:
                continue
