pyproject.toml
# pyproject.toml
[tool.poetry]
name = "keyguard"
version = "0.1.0"
description = "A keylogger prevention tool for Windows and Linux"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyQt6 = "^6.4.0"
psutil = "^5.9.0"
pyuac = "^0.0.3"
python-xlib = {version = "^0.33", platform = "linux"}
win32security = {version = "^305.1", platform = "win32"}

[tool.poetry.scripts]
keyguard = "keyguard.cli:main"
